Adjust camera orientation and person pose in simulation
Updated the camera joint's roll-pitch-yaw in camera.xacro for camera to look slightly up (because before I didn't see the head of the human entity). Changed the person_standing model's pose in cbrn_world.world to stand instead of laying down to test if Media Pipe Pose is correctly identifying the person.